+++ src/databricks/azext_databricks/commands.py
@@ -14,10 +14,14 @@ from ._validators import validate_workspace_values
 def load_command_table(self, _):
 
     from ._client_factory import cf_workspaces
+
+    with self.command_group('databricks', is_preview=True):
+        pass
+
     databricks_workspaces = CliCommandType(
         operations_tmpl='azext_databricks.vendored_sdks.databricks.operations._workspaces_operations#WorkspacesOperations.{}',
         client_factory=cf_workspaces)
-    with self.command_group('databricks workspace', databricks_workspaces, client_factory=cf_workspaces, is_experimental=True) as g:
+    with self.command_group('databricks workspace', databricks_workspaces, client_factory=cf_workspaces) as g:
         g.custom_command('create', 'create_databricks_workspace', validator=validate_workspace_values, supports_no_wait=True)
         g.custom_command('update', 'update_databricks_workspace', supports_no_wait=True)
         g.custom_command('delete', 'delete_databricks_workspace', supports_no_wait=True, confirmation=True)