+++ src/spring-cloud/azext_spring_cloud/vendored_sdks/appplatform/_app_platform_management_client.py
@@ -20,6 +20,7 @@ from .operations import CertificatesOperations
 from .operations import CustomDomainsOperations
 from .operations import DeploymentsOperations
 from .operations import Operations
+from .operations import SkuOperations
 from . import models
 
 
@@ -43,6 +44,8 @@ class AppPlatformManagementClient(SDKClient):
     :vartype deployments: azure.mgmt.appplatform.operations.DeploymentsOperations
     :ivar operations: Operations operations
     :vartype operations: azure.mgmt.appplatform.operations.Operations
+    :ivar sku: Sku operations
+    :vartype sku: azure.mgmt.appplatform.operations.SkuOperations
 
     :param credentials: Credentials needed for the client to connect to Azure.
     :type credentials: :mod:`A msrestazure Credentials
@@ -78,4 +81,6 @@ class AppPlatformManagementClient(SDKClient):
         self.deployments = DeploymentsOperations(
             self._client, self.config, self._serialize, self._deserialize)
         self.operations = Operations(
+            self._client, self.config, self._serialize, self._deserialize)
+        self.sku = SkuOperations(
             self._client, self.config, self._serialize, self._deserialize)